+++ frb_codegen/src/library/codegen/parser/hir/naive_flat/transformer/mod.rs
@@ -1,2 +1,2 @@
-pub(crate) mod module_filter_transformer;
+pub(crate) mod filter_transformer;
 pub(crate) mod move_third_party_override_transformer;
